requirements.txt
asgiref==3.8.1
astroid==3.2.4
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dill==0.3.8
Django==5.0.7
django-keycloak-auth==1.0.0
djangorestframework==3.15.2
idna==3.7
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.8.0
pylint==3.2.6
requests==2.32.3
sqlparse==0.5.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.44.0
